update ServiceDiscovery instance information handling
This carries a number of changes:

Adds the instance name to InstanceInformation, with this change,
the instance info (ip address and port) is now added when creating the service instead of calling `add_service_info`.

It also improves the filter function in the resource record manager by introducing a filter struct.

Cached resources are no longer replied to queries, since this is against the mDNS protocol